requirements.txt
requests==2.18.4
flask==0.12.2
flask_script==2.0.6
flask_restful==0.3.6
python_dotenv==0.8.2
apscheduler==3.5.1
tinydb==3.8.1
bs4==0.0.1
pytz==2016.10
python-dateutil==2.6.0